requirements.txt
requests
flask
python-dotenv
flask_sqlalchemy
flask_cors
sqlalchemy